app/requirements.txt
faiss-cpu==1.5.3
ipython==8.12.3
kokoro_onnx==0.4.9
langchain_text_splitters==1.1.0
numpy==2.3.5
ollama==0.6.1
pandas==2.3.3
PyPDF2==3.0.1
scipy==1.16.3
sentence_transformers==5.2.0
streamlit==1.49.1
